sched/tune: Add support for dynamic 'crucial'
This could be highly useful for processes that needs a lot of processing power wherein we could trigger this function that will allow the scheduler to prioritize and assign them to a highly capable cpu. This will greatly speed up the processes and finish up the tasks a lot quicker.

Signed-off-by: ederekun <sedrickvince@gmail.com>