dynamic_stune: Replace extend timer with an update atomic
Having mod_timer in multiple threads is not a great idea, let's again make use of atomics to replace it. Extend the update atomic to also accomodate for INPUT and ensure that we offload the timeout refresh to the GPU and timer itself once the GPU *maintains* the maximum frequency (avoid short bursts from getting an update). Move the update call after filtered by userspace's and QoS limits to reflect the registered GPU freq.

Signed-off-by: ederekun <sedrickvince@gmail.com>