sound: qdsp6v2: Adapt DTS Eagle ASM and ADM methodologies to the new instance ID API
As the driver was reverted prior to the API, there was a need to adapt the structures to abide with the new instance ID and header packing standard. Create a duplicate structure for asm_stream_cmd_get based on asm_stream_cmd_set given that it is the format now recognized by the apr. Get rid of unnecessary parts as much as possible.

Signed-off-by: ederekun <sedrickvince@gmail.com>