sound: qdsp6v2: Force DTS HPX COPP for LEGACY_PCM ADM, set hwacc effects ASM topology to HPX master, and compressed playback audio effects topology to HPX plus
Insist on using the memory regions for DTS by redirecting ASM and ADM towards them. Only redirect for LEGACY_PCM and while ADM path is configured for playback. Also modify soft volume handlers to be synced up as it is unlikely that userspace will call to topology switch soft volume instance due to the removal of DTS HPX, this should prevent inconsistencies with volume level throughout by making use of the values sent to SA+.

Signed-off-by: ederekun <sedrickvince@gmail.com>